Fixed a bug with the correct pointer passed as the RAM address
The CPU was originally designed to have asynchronous RAM reads. This meant
that whenever DP was changed, the RAM data output value would update.
However, the RAM is now fully synchronous (to take advantage of FPGA block
RAM), which meant that it now took TWO clock cycles for the value to
update (one to update the DP and another to update the RAM data output).

To solve this, next_dp is now attached to the memory address.